requirements.txt
Django>=4.2
djangorestframework
drf-yasg
Faker
python-decouple
